refactor: extract the embeddings lifecycle from IndexManager
Refs #1157, refs #888, refs #736
